CommitTechnical Assignment
1. Develop a solution based upon the requirements and details below.
2. When completed, check the code into your repository of choice (GitHub tends to be the most popular).
3. Submit the link to the repository back to your recruiter.

Application Requirements
1. Create a .NET console application that:
a. Prompts user for a file name.
b. Searches directory for a .csv file with that name.
c. Parse CSV File
i. If the file exists, iterate through each row and validate that the email address is valid syntax.
1. Track separate lists for valid and invalid email addresses. (e.g. test@gdcit.com is valid,  test@gmail is not valid)
ii. If the file does not exist, return an error message to the user.
d. Output the lists of valid and invalid email addresses found.

Other Details
1. CSV File Structure
a. Fields in the CSV file should be First Name, Last Name, Email Address
Add project files.